cipher: make encryption more secure
Previous version of the cipher used only password and index of a chunk to
encrypt the chunk. So if someone knew a message and its code, then they can
decrypt any other code of the same length. Current version solves this problem
by additionally using hash of processed prefix and hash of the whole message.